services/dft_predict/requirements.txt
simms
codex-africanus[complete]
astropy
numpy
scipy
astLib
python-casacore
future
astro-kittens
matplotlib
jupyter